% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedr_dependent.R
\name{assert_that_bedtools_binary_available}
\alias{assert_that_bedtools_binary_available}
\title{Check for Bedtools Binary}
\usage{
assert_that_bedtools_binary_available(
  directory_containing_binary_search = NULL,
  silent = FALSE
)
}
\arguments{
\item{directory_containing_binary_search}{directory to search for binaries within (will be preppended onto PATH temporarily). If NULL, will search through PATH.}

\item{silent}{supress all output headed to stderr()}
}
\value{
(invisible) TRUE if bedtools binary was found, throws error if it was not
}
\description{
Check for Bedtools Binary
}
\examples{
assert_that_bedtools_binary_available()
}
