% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedfile_tests.R
\name{bed_chromosomes_match_genome_file}
\alias{bed_chromosomes_match_genome_file}
\title{Bedfile chromosome correctness}
\usage{
bed_chromosomes_match_genome_file(bedfile_path, genome, warn = TRUE)
}
\arguments{
\item{bedfile_path}{path to bed file}

\item{genome}{path to bedtools genome file. To make one, see: https://www.biostars.org/p/70795/}

\item{warn}{Print warning messages if bed contains chromosomes not in the genomefile (logical)}
}
\value{
TRUE if all chromosomes in BEDfile are also present in the genome file
}
\description{
Bedfile chromosome correctness
}
