% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtoolsr_extension.R
\name{bed_intervals_all_occur_within_regions}
\alias{bed_intervals_all_occur_within_regions}
\title{Do All Intervals Lie Within Regions}
\usage{
bed_intervals_all_occur_within_regions(input_bed, regions_bed)
}
\arguments{
\item{input_bed}{path to a bedfile whose intervals you want to test (string)}

\item{regions_bed}{path to a bedfile defining regions you expect will encapsulate input_bed intervals (string)}
}
\value{
TRUE if all input_bed intervals are encapsulated in regions_bed intervals. Otherwise returns FALSE. (logical)
}
\description{
Takes two bedfiles and tests whether intervals in the first are ALL encapsulated by intervals in the second.
Useful function for checking that restriction of some feature to particular bed intervals worked as expected
}
\details{
May also work with other bedtools compatible input formats (e.g. vcfs) although this functionality is untested. Use at your own risk.
}
