% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtoolsr_extension.R
\name{bed_intervals_overlap}
\alias{bed_intervals_overlap}
\title{Do intervals within a bedfile overlap}
\usage{
bed_intervals_overlap(
  bedfile_path,
  force_strandedness = FALSE,
  header = FALSE,
  warn = TRUE,
  verbose = TRUE,
  cleanup_temp = TRUE
)
}
\arguments{
\item{bedfile_path}{path to bedfile (string)}

\item{force_strandedness}{should we consider it impossible for intervals on one strand to overlap with those on the other (logical)}

\item{header}{does my bedfile have a header (logical)}

\item{warn}{print warnings to stderr  (logical)}

\item{verbose}{verbose mode (logical)}

\item{cleanup_temp}{delete temporary files after use (logical)}
}
\value{
TRUE if redundancy is present, FALSE if each region is completely distinct (logical)
}
\description{
Checks whether intervals WITHIN a bedfile overlap. If so, it is usually a good idea to run bedtools merge
}
