% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedfile_tests.R
\name{bed_is_sorted}
\alias{bed_is_sorted}
\title{Is Bedfile Sorted?}
\usage{
bed_is_sorted(bedfile_path, genome, header = NULL, warn = TRUE)
}
\arguments{
\item{bedfile_path}{path to bed file}

\item{genome}{path to bedtools genome file. To make one, see: https://www.biostars.org/p/70795/}

\item{warn}{Print warning messages if bed is unsorted (logical)}
}
\value{
TRUE if sorted as per genome file, FALSE if not (logical)
}
\description{
Is Bedfile Sorted?
}
