% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtoolsr_extension.R
\name{bed_randomly_shuffle_intervals}
\alias{bed_randomly_shuffle_intervals}
\title{Randomly Shuffle Intervals}
\usage{
bed_randomly_shuffle_intervals(
  input,
  regions,
  genome,
  iterations = 1L,
  outdir,
  force = FALSE,
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{input}{path to input bed containing intervals to re-shuffle (string)}

\item{regions}{path to bed file representing regions within which input intervals can be reshuffled (string)}

\item{genome}{path to genome file (string)}

\item{iterations}{number of times to independently reshuffle the bedfile}

\item{outdir}{directory to save output}

\item{force}{force overwrite if directory exists}

\item{seed}{seed for random number generation, used for reproducibility. When seed is supplied & iterations > 1, the nth iteration is seeded with \strong{seed + (n-1)} (numeric)}

\item{verbose}{verbose mode}
}
\value{
path to directory containing the input files (string)
}
\description{
Takes an input bed and reshuffles intervals randomly throughout the genome (restricted to regions defined by \strong{regions} option).
The size of each interval is preserved, its just moved randomly to a new location.
This is function is a useful first step if looking to see if particular regions are enriched for some feature because it allow you to examine such features (e.g. variant density) in the original intervals vs randomly reshuffled intervals.
}
