% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedstats.R
\name{bed_total_region_size}
\alias{bed_total_region_size}
\title{bed_total_region_size}
\usage{
bed_total_region_size(data)
}
\arguments{
\item{data}{a bed-style dataframe where 2nd column is start and third is stop}
}
\value{
summed length of all intervals
}
\description{
R equivalent of:
awk '{sum += $3-$2} END{print sum}'

WARING: Any redundant segments of your bed-style dataframe will be counted multiple times.
}
