% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtoolsr_extension.R
\name{bed_variant_densities}
\alias{bed_variant_densities}
\title{bed_variant_densities}
\usage{
bed_variant_densities(
  bedfiles,
  variantfile,
  genomefile,
  scaling = "kb",
  bedfilepaths_to_names = function(x) {     return(x) }
)
}
\arguments{
\item{bedfiles}{path to bed file (string)}

\item{variantfile}{path to bedtools-compatible variant file, e.g. bed/vcf (string)}

\item{genomefile}{path to genome file (string)}

\item{scaling}{the unit of variant density. For example, the default 'kb' will force the function to return variants/kilobase. Options include 'b'  (base), 'kb' (kilobase) and 'mb' (megabase) (string)}

\item{bedfilepaths_to_names}{a function that takes a string (the bedfile path) and outputs a string that will be used as the 'name' of the returned value. Default uses the full filename (function)}
}
\value{
named vector where values are variants per base, and names are derived from each bedfile path supplied (double)
}
\description{
Takes n bed files and returns the average variant density (variants per bp) for each
}
\examples{
\dontrun{
bed_variant_densities(
bedfiles = c(input_bed,regions.bed),
variantfile = variants.bed,
genomefile = genome,
bedfilepaths_to_names = function(x) {
utilitybelt::path_process_vectorized(paths = x, extract_basename = TRUE, remove_extension = TRUE)
})
}
}
