% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtoolsr_extension.R
\name{bed_variant_densities_compare}
\alias{bed_variant_densities_compare}
\title{bed_variant_densities_compare}
\usage{
bed_variant_densities_compare(
  bedfiles,
  variantfile,
  genomefile,
  scaling = "kb",
  bedfilepaths_to_names = function(x) {     return(x) },
  comparator
)
}
\arguments{
\item{bedfiles}{path to bed file (string)}

\item{variantfile}{path to bedtools-compatible variant file, e.g. bed/vcf (string)}

\item{genomefile}{path to genome file (string)}

\item{scaling}{the unit of variant density. For example, the default 'kb' will force the function to return variants/kilobase. Options include 'b'  (base), 'kb' (kilobase) and 'mb' (megabase) (string)}

\item{bedfilepaths_to_names}{a function that takes a string (the bedfile path) and outputs a string that will be used as the 'name' of the returned value. Default uses the full filename (function)}

\item{comparator}{a function that takes two arguments. 1. the variant density of bedfile. 2. the variant density of bedfile2. The function should returns some metric of the difference. For example, magrittr::subtract would subtract the second argument from the first. (function)}
}
\value{
result of comparator_function (variable -- usually numeric but could be boolean or anything else)
}
\description{
Compares variant density of two bedfiles with a function of your choosing
}
