% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtoolsr_extension.R
\name{bed_variant_densities_compare_directory_of_beds_to_single_ref}
\alias{bed_variant_densities_compare_directory_of_beds_to_single_ref}
\title{bed_variant_densities_compare_directory_of_beds_to_single_ref}
\usage{
bed_variant_densities_compare_directory_of_beds_to_single_ref(
  directory_containing_bed,
  reference_bedfile,
  variantfile,
  genomefile,
  scaling = "kb",
  bedfilepaths_to_names = function(x) {     return(x) },
  comparator = magrittr::subtract,
  output_checking_function = is.numeric
)
}
\arguments{
\item{directory_containing_bed}{A directory containing exclusively bedfiles}

\item{reference_bedfile}{path to a bedfile whose variant density is going to be compared to each variant density of the bedfiles in 'directory_containing_bed' (string)}

\item{variantfile}{path to bedtools-compatible variant file, e.g. bed/vcf (string)}

\item{genomefile}{path to genome file (string)}

\item{scaling}{the unit of variant density. For example, the default 'kb' will force the function to return variants/kilobase. Options include 'b'  (base), 'kb' (kilobase) and 'mb' (megabase) (string)}

\item{bedfilepaths_to_names}{a function that takes a string (the bedfile path) and outputs a string that will be used as the 'name' of the returned value. Default uses the full filename (function)}

\item{comparator}{a function that takes two arguments. 1. the variant density of bedfile. 2. the variant density of bedfile2. The function should returns some metric of the difference. For example, magrittr::subtract would subtract the second argument from the first. (function)}

\item{output_checking_function}{a function that takes the output vector and returns TRUE if it is as expected and FALSE if it is not. Output of this function is fed into an assertion. (function w/ logi)}
}
\value{
vector of densities_compared_to_ref (usually numeric but could vary depending on comparator)
}
\description{
bed_variant_densities_compare_directory_of_beds_to_single_ref
}
