% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtoolsr_extension.R
\name{bed_variant_densities_compare_subtract}
\alias{bed_variant_densities_compare_subtract}
\title{bed_variant_densities_compare_divide}
\usage{
bed_variant_densities_compare_subtract(
  bedfiles,
  variantfile,
  genomefile,
  scaling = "kb",
  bedfilepaths_to_names = function(x) {     return(x) }
)
}
\arguments{
\item{bedfiles}{path to bed file (string)}

\item{variantfile}{path to bedtools-compatible variant file, e.g. bed/vcf (string)}

\item{genomefile}{path to genome file (string)}

\item{scaling}{the unit of variant density. For example, the default 'kb' will force the function to return variants/kilobase. Options include 'b'  (base), 'kb' (kilobase) and 'mb' (megabase) (string)}

\item{bedfilepaths_to_names}{a function that takes a string (the bedfile path) and outputs a string that will be used as the 'name' of the returned value. Default uses the full filename (function)}
}
\value{
result of comparator_function (variable -- usually numeric but could be boolean or anything else)
}
\description{
Compares variant density of two bedfiles by comparing the absolute difference between the two density values
Variant density is given in variants/base
}
