% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtoolsr_extension.R
\name{bed_variant_density}
\alias{bed_variant_density}
\title{Density of variants in bed file}
\usage{
bed_variant_density(bedfile, variantfile, genomefile, scaling = "kb")
}
\arguments{
\item{bedfile}{path to bed file (string)}

\item{variantfile}{path to bedtools-compatible variant file, e.g. bed/vcf (string)}

\item{genomefile}{path to genome file (string)}

\item{scaling}{the unit of variant density. For example, the default 'kb' will force the function to return variants/kilobase. Options include 'b'  (base), 'kb' (kilobase) and 'mb' (megabase) (string)}
}
\value{
variants per base / kilobase / megabase depending on 'scaling' (number)
}
\description{
Density of variants in bed file
}
