% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtoolsr_extension.R
\name{bed_variants_per_interval}
\alias{bed_variants_per_interval}
\title{bed_variants_per_interval}
\usage{
bed_variants_per_interval(bedfile, variantfile, genomefile, outfile = NULL)
}
\arguments{
\item{bedfile}{path to bed file (string)}

\item{variantfile}{path to bedtools-compatible variant file, e.g. bed/vcf (string)}

\item{genomefile}{path to genome file (string)}

\item{outfile}{file to write results to. If not supplied, will return an R dataframe instead(string)}
}
\value{
a file/dataframe (depends on whether or not outfile is supplied) with a bed format layout containing an extra column with thenumber of
}
\description{
bed_variants_per_interval
}
