% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedfile_tests.R
\name{file_has_header}
\alias{file_has_header}
\title{Guess if file has header}
\usage{
file_has_header(filepath)
}
\arguments{
\item{filepath}{a filepath}
}
\value{

}
\description{
uses data.table::fread to guess if file has header. TODO:::TESTTT
}
