% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedfile_tests.R
\name{file_is_bed}
\alias{file_is_bed}
\title{Check if file is bed}
\usage{
file_is_bed(filepath)
}
\arguments{
\item{filepath}{path to putative bedfile}

\item{genomefile}{path to bedtools genome file. To make one, see: https://www.biostars.org/p/70795/}
}
\value{
(logical)
}
\description{
Check columns 2 and 3 are integers
}
