% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedfile_tests.R
\name{file_is_valid_bed}
\alias{file_is_valid_bed}
\title{Is file a valid, sorted bed
Is file a valid, sorted bed}
\usage{
file_is_valid_bed(filepath, genomefile, warn = TRUE)
}
\arguments{
\item{filepath}{path to putative bedfile (string)}

\item{genomefile}{path to bedtools genome file. To make one, see: https://www.biostars.org/p/70795/ (string)}
}
\value{
(logical)
}
\description{
Is file a valid, sorted bed
Is file a valid, sorted bed
}
