% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedfile_tests.R
\name{file_is_valid_bed_vectorized}
\alias{file_is_valid_bed_vectorized}
\title{Are files real, valid beds}
\usage{
file_is_valid_bed_vectorized(filepaths, genomefile, warn = TRUE)
}
\arguments{
\item{filepaths}{vector of paths to putative bedfiles (character)}

\item{genomefile}{path to bedtools genome file. To make one, see: https://www.biostars.org/p/70795/ (string)}
}
\value{

}
\description{
Are files real, valid beds
}
