% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtoolsr_extension.R
\name{variant_density_compare_all_to_ref}
\alias{variant_density_compare_all_to_ref}
\title{Compare several variant densities to single reference}
\usage{
variant_density_compare_all_to_ref(
  variant_densities,
  reference_variant_density,
  comparator,
  output_checking_function = is.numeric
)
}
\arguments{
\item{variant_densities}{vector of variant densities}

\item{reference_variant_density}{a single variant density you want to compare#'}

\item{comparator}{a function that takes two arguments. 1. the variant density of bedfile. 2. the variant density of bedfile2. The function should returns some metric of the difference. For example, magrittr::subtract would subtract the second argument from the first. (function)}

\item{output_checking_function}{a function that takes the output vector and returns TRUE if it is as expected and FALSE if it is not. Output of this function is fed into an assertion. (function w/ logi)}
}
\value{
for each variant density metric supplied, this function returns the output of a comparator function (output vector is equal in length to input variant_densities vector) (vector of any_value_type)
}
\description{
Takes a bunch of raw variant densities + a single 'reference variant density' and applies some comparator function \strong{ f(raw_variant_density, reference_variant_density) = some_metric}
This comparator function is run separately for each raw_variant_density, resulting in a vector of outputs equal to length(variant_densities).
}
